Add device-log builder tests
